Use bundled terraform-ls
This converts the extensiom to use the bundled terraform-ls binary instead of one downloaded at runtime. This effectively changes the extension to not download the terraform-ls binary and instead rely on one being present in the `bin` directory inside the extension path.

It also moves @hashicorp/js-releases to a development dependency, which then allows openpgp, semver, and yauzl to be removed. These dependencies are no longer needed in the extension as we no longer download the terraform-ls bianry at runtime.

We still will use @hashicorp/js-releases to download the terraform-ls binary at development and package time.